Renamed clone to cloner, re-enabled test
